requirements.txt
Django>=4.0
Pillow>=9.1.0
requests>=2.2
polyline>=1.4.0
python_tsp>=0.2.1
folium>=0.12
reportlab>=3.6.9